chore(wip): start applying new layout to filters
