+++ packages/frontend/src/features/semanticViewer/components/FiltersModal/getOperatorString.ts
